Refactored to a more idiomatic KSP implementation (#16)
The original POC of this project modelled Anvil a little too closely in
a way that is just unecessary with KSP, i.e. having all the code gen in
a single SymbolProcessor.

This PR breaks out all the codegen into their own SymbolProcessor to
take advantage of KSP built-in multi-round support and greatly
simplifies the implementation without having to do any weird caching of
generated hints or the like.

This PR also moves the Circuit compilation stuff into its own module
that can be applied seperately